Support landscapist module for kmp
