+++ landscapist/src/androidMain/AndroidManifest.xml
