+++ landscapist/src/commonMain/kotlin/com/skydoves/landscapist/StableHolder.kt
